# ENH: Use pyarrow.compute for unique, dropna