--- a/pandas/tests/base/test_unique.py
+++ b/pandas/tests/base/test_unique.py
@@ -1,5 +1,3 @@
-from contextlib import nullcontext
-
 import numpy as np
 import pytest
 
@@ -14,18 +12,12 @@ from pandas.core.api import NumericIndex
 from pandas.tests.base.common import allow_na_ops
 
 
-def maybe_perf_warn(using_pyarrow):
-    if using_pyarrow:
-        return tm.assert_produces_warning(PerformanceWarning, match="Falling back")
-    else:
-        return nullcontext()
-
-
 def test_unique(index_or_series_obj):
     obj = index_or_series_obj
     obj = np.repeat(obj, range(1, len(obj) + 1))
-    with maybe_perf_warn(
-        pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]"
+    with tm.maybe_produces_warning(
+        PerformanceWarning,
+        pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]",
     ):
         result = obj.unique()
 
@@ -65,8 +57,9 @@ def test_unique_null(null_obj, index_or_series_obj):
     klass = type(obj)
     repeated_values = np.repeat(values, range(1, len(values) + 1))
     obj = klass(repeated_values, dtype=obj.dtype)
-    with maybe_perf_warn(
-        pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]"
+    with tm.maybe_produces_warning(
+        PerformanceWarning,
+        pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]",
     ):
         result = obj.unique()
 
@@ -93,8 +86,9 @@ def test_unique_null(null_obj, index_or_series_obj):
 def test_nunique(index_or_series_obj):
     obj = index_or_series_obj
     obj = np.repeat(obj, range(1, len(obj) + 1))
-    with maybe_perf_warn(
-        pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]"
+    with tm.maybe_produces_warning(
+        PerformanceWarning,
+        pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]",
     ):
         expected = len(obj.unique())
     assert obj.nunique(dropna=False) == expected
@@ -120,16 +114,19 @@ def test_nunique_null(null_obj, index_or_series_obj):
         assert obj.nunique() == len(obj.categories)
         assert obj.nunique(dropna=False) == len(obj.categories) + 1
     else:
-        with maybe_perf_warn(
-            pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]"
+        with tm.maybe_produces_warning(
+            PerformanceWarning,
+            pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]",
         ):
             num_unique_values = len(obj.unique())
-        with maybe_perf_warn(
-            pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]"
+        with tm.maybe_produces_warning(
+            PerformanceWarning,
+            pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]",
         ):
             assert obj.nunique() == max(0, num_unique_values - 1)
-        with maybe_perf_warn(
-            pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]"
+        with tm.maybe_produces_warning(
+            PerformanceWarning,
+            pa_version_under2p0 and str(index_or_series_obj.dtype) == "string[pyarrow]",
         ):
             assert obj.nunique(dropna=False) == max(0, num_unique_values)