--- a/pandas/tests/extension/test_string.py
+++ b/pandas/tests/extension/test_string.py
@@ -13,7 +13,6 @@ classes (if they are relevant for the extension interface for all dtypes), or
 be added to the array-specific tests in `pandas/tests/arrays/`.
 
 """
-from contextlib import nullcontext
 import string
 
 import numpy as np
@@ -27,13 +26,6 @@ import pandas._testing as tm
 from pandas.core.arrays import ArrowStringArray
 from pandas.core.arrays.string_ import StringDtype
 from pandas.tests.extension import base
-
-
-def maybe_perf_warn(using_pyarrow):
-    if using_pyarrow:
-        return tm.assert_produces_warning(PerformanceWarning, match="Falling back")
-    else:
-        return nullcontext()
 
 
 def split_array(arr):
@@ -152,8 +144,9 @@ class TestIndex(base.BaseIndexTests):
 
 class TestMissing(base.BaseMissingTests):
     def test_dropna_array(self, data_missing):
-        with maybe_perf_warn(
-            pa_version_under6p0 and data_missing.dtype.storage == "pyarrow"
+        with tm.maybe_produces_warning(
+            PerformanceWarning,
+            pa_version_under6p0 and data_missing.dtype.storage == "pyarrow",
         ):
             result = data_missing.dropna()
         expected = data_missing[[1]]