… pandas/tests/strings/test_find_replace.py
… pandas/tests/strings/test_strings.py