@simonjayhawkins 2022-04-10
> my understanding of #42613 is that ...

more of the discussion was actually in #42597