@jreback 2022-04-10
> Ah thanks, I wasn't aware of this discussion.
> 
> My prior, related PR's so far have just been moving existing methods, so no fallback behavior should have been introduced I think.
> 
> For this PR, I will just remove the super calls for now and reintroduce them later if we decide on a different fallback policy

comments on the other issues is we can show a PerformanceWarning if we are falling back. But let's do that as a pre-cursor