@jreback 2022-04-26
actually if you can do that move in this PR and I think this also needs a whatsnew note. ping on green.