@mroeschke 2022-04-26
Here are the perf differences (I think I see what you mean about the dropna dispatch)

PR
```
In [1]: import pyarrow as pa
   ...: pa.__version__
   ...:
Out[1]: '7.0.0'

In [2]: ser = pd.Series(["1", np.nan] * 100_000, dtype="string[pyarrow]")

In [4]: %timeit ser.unique()
   ...:
1.3 ms ± 6.59 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [5]: %timeit ser.dropna()

2.27 ms ± 46.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

main
```
In [1]: ser = pd.Series(["1", np.nan] * 100_000, dtype="string[pyarrow]")
   ...:

In [2]: %timeit ser.unique()
   ...:
24.4 ms ± 181 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [3]: %timeit ser.dropna()
2.27 ms ± 14.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```